feat: move util/verified-content to root shared folder
