+++ shared/util/verified-content.ts
